requirements.txt
Django==3.2
pytest-django
psycopg2-binary
requests
djangorestframework
djangorestframework-simplejwt
pylint
django-environ
httpie
pylint
django-environ
pylint-django
django-celery-results
django-celery-beat
pytz
PyJWT==v2.3.0
celery
django-crontab
